LayoutTests/imported/w3c/web-platform-tests/html/semantics/tabular-data/the-table-element/tFoot-expected.txt


FAIL tFoot tests assert_equals: expected Element node <tbody id="tbody2"></tbody> but got Element node <thead id="thead"></thead>
